Prevent insertion of additional linebreak in message body
The use of bufio.Scanner meant that linebreaks were handled as part of
the scanner internals. By always appending a linebreak after each call
to scanner.Scan(), it could happen that the final chunk of our input
stream did not end with a newline, causing us to add one incorrectly.

As a result of this, the original email message was incorrectly modified
resulting in, for example, DKIM signatures becoming invalid.

By replacing the scanner with the slightly more low-level bufio.Reader,
we preserve linebreaks in the input we are handling. This means we don't
unconditionally add a linebreak ourselves, avoiding the problem
altogether.

Fixes #5, closes #6