Intelligently guess correct line endings
SMTP specifies \r\n for line endings, but sendmail is not actually an
SMTP transaction. While not clearly documented (that I could find),
sendmail being a unix command-line application it likely follows the
convention of using the system's native line-breaks (\r\n on Windows, \n
on Linux, *BSD, etc)

However, naively using the native line endings could result in mixed
use, especially when debugging issues where you're manually piping in
input.

To deal with this, adopt a similar behavior of postfix uses, but basing
the decision of what line ending to use off of the line ending used on
the first line of input.